White list something.

+++ application/src/com.phdroid/smsb/storage/dao/SmsMessageEntry.java
@@ -18,7 +18,7 @@ public class SmsMessageEntry extends SmsPojo {
 	public static final String ACTION = "action";
 
 	private long id;
-	private int senderId;
+	private long senderId;
 	private String message;
 	private long received;
 	private boolean read;
@@ -65,11 +65,11 @@ public class SmsMessageEntry extends SmsPojo {
 		return id;
 	}
 
-	public int getSenderId() {
+	public long getSenderId() {
 		return senderId;
 	}
 
-	public void setSenderId(int senderId) {
+	public void setSenderId(long senderId) {
 		this.senderId = senderId;
 	}
